Add Volume slider for altering volume
Adds a volume slider. Now users can increase/decrease the volume
of the video, as a part of manipulations. Live change can be observed on
client side. For server side, the changes (made by running a ffmpeg
command in background) are reflected as soon as the user previews the
video.

Bug: T336597
Change-Id: Ib3a671bea4766a1a1962db7495f362e4115cce31